pyproject.toml
[project]
name = "profile-matcher-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.16",
    "fastapi[standard]>=0.116.1",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pydantic-settings>=2.10.1",
    "supabase>=2.17.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "requests>=2.32.4",
    "aiohttp>=3.12.14",
]
